Change Interface/InterfaceMeta and class generation
Interface now directly metaclasses InterfaceMeta, making all of the
protocol Interface's not have to.  Furthermore, the various class
generation (Proxy, etc.) from interfaces is moved into InterfaceMeta
such that the property acts as a class property, and can be used by
client.Display to create the correct class with all of the requisite
methods.