Get rid of __del__ methods
Move all of the logic in __del__ method out into ffi.gc methods.  This
should be a bit more robust and the notion of keeping things alive is
correct. The one corner case, I think, is when you start deleting
without running the funalizers (as may be the case with PyPy), and then
you run something like disconnect, which may disconnect before the
other finalizers run.  Maybe I can have the .disconnect just delete the
cdata, and trigger its finalizer.