Fix up finalizers to work better with PyPy
In PyPy, there is no guarantee when the finalizers will be run, so the
cleanup defined by `ffi.gc()` may not run immediately.  Add some
explicit gc.collect() calls in the tests that need destructor callbacks
to fire.

Also, for the destructors to fire in the correct order, there needs to
be a direct chain of refernce ownership that frees all other objects
before freeing a display. We need to:
  1. Make the freeing more robust, e.g. so a global won't try to destroy
     its pointer after the display has been destroyed. In case the user
     creates a reference loop or frees things in the wrong order, we
     don't want to be segfaulting.
  2. Get rid of reference loops, namely `self.handle =
     ffi.new_handle(self)`, since the handle keeps a reference to
     `self`. This has been addressed everywhere except for
     proxy/resource objects.