3.2 is dropped, no current pypy3 version supported
The most recent pypy3 version is too old to have the cffi features that
we need, so remove this from the list of Travis tests.  There is a
current version on PyPy that works (>=5.0), so leave that in the tests,
but the tests here will keep failing because the Travis version is too
old.