Fix trying to release disconnected proxies
Change the destruction of the proxies such that they will not be
destroyed if the display they are associated with has already been
disconnected.  There is some other reason where the weakset of proxy
elements does not contain a live reference to the proxy object, as the
handling there ought to destroy the proxy.