Drop Python 3.5 support
Remove support for Python 3.5, which was already dubios because of the
IntEnum use in the protocols.  Make the change to the types that are
declared as comments to use the newer syntax.  This also unlocks the
ability to use dataclasses (in 3.6 via a backport), f strings, and make
heavier use of IntEnum's.