Change proxy and resource classes to be generated
Rather than using dynamically created proxy and resource class types,
generated by parsing the needed elements out of the interface, have the
generated protocols directly define the proxy and resource classes.
This removes the need to do some of the metaprogramming in the interface
to create these classes and simplifies using the generated instances as
types.