Fix problem with pip installing
There is a problem trying to pip install version 0.2.0 because of the
sharing of the ArgumentType enum class, in particular, that it is
imported out of the protocol_core module for use in the scanner.
Because the ffi modules have not been built at install time, we are not
able to perform this import.  Move the definition to be in the scanner
and setup the protocol_core to import this instead.