Change installation to not trigger rebuild when pip installing
The current package installation will trigger a rebuild of the protocols
when the package is built and again when the package is pip installed.
This changes the package to be built only at package creation time,
which should guarantee that the protocols are installed.  If this needs
to be more dynamic per machine, this can be changed to instead use the
system version.  The location is resolved using pkg-config, so this
should be ok to do.  Currently, it will do a full build on package build
via sdist. To change to packaging at install time, we would need to
change the sdist command to an install command.  We would also need to
change the manifest to not package the protocols that may have been
built for local development.